modules/m5-clipping/requirements.txt
# M5 Clipping Service Dependencies
# Python 3.11+

# Flask API
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Google Cloud
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-api-core==2.14.0

# Video Processing
ffmpeg-python==0.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
